add getDirectory and addDirectory + refactoring code

+++ src/main/java/org/example/cloudstorage/service/FileService.java
@@ -5,10 +5,7 @@ import io.minio.errors.*;
 import io.minio.messages.Item;
 import lombok.RequiredArgsConstructor;
 import org.example.cloudstorage.dto.FileInfoDto;
-import org.example.cloudstorage.exception.AppException;
-import org.example.cloudstorage.exception.InvalidInputDataException;
-import org.example.cloudstorage.exception.ObjectNotFoundException;
-import org.example.cloudstorage.exception.UserIsNotAuthenticated;
+import org.example.cloudstorage.exception.*;
 import org.example.cloudstorage.util.FilePathUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.InputStreamResource;
@@ -18,10 +15,8 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
+import java.io.ByteArrayInputStream;
 import java.io.InputStream;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,12 +39,25 @@ public class FileService {
     }
 
     public FileInfoDto upload(MultipartFile file, String path){
+        if(!FilePathUtil.isValidPath(path) || FilePathUtil.getType(path).equals("DIRECTORY")){
+            throw new InvalidInputDataException("Invalid path");
+        }
+
+        if(isPathToResourceExists(path)){
+            throw new ObjectAlreadyExistsException("already exists");
+        }
+
+        if(!isPathToResourceExists(FilePathUtil.getResourcePath(path))){
+            throw new ObjectNotFoundException("Parent folder didn't exist");
+        }
+
         try {
-            minioClient.putObject(PutObjectArgs.builder().bucket(bucketName).object(getUserPrefix() + file.getName()).stream(file.getInputStream(), -1, 10485760 ).build());
-        } catch (Exception e) {
+            minioClient.putObject(PutObjectArgs.builder().bucket(bucketName).object(getUserPrefix() + file.getName()).stream(file.getInputStream(), -1, 10485760).build());
+            return getInfo(getUserPrefix() + path + file.getName());
+        }
+        catch (Exception e) {
             throw new RuntimeException();
         }
-        return null;
     }
 
     public FileInfoDto getInfo(String path) {
@@ -60,7 +68,7 @@ public class FileService {
             StatObjectResponse response =
                     minioClient.statObject(
                             StatObjectArgs.builder().bucket(bucketName).object(getUserPrefix() + path).build());
-            return new FileInfoDto(FilePathUtil.getResourceName(path), FilePathUtil.getPath(path), response.size(), FilePathUtil.getType(path));
+            return new FileInfoDto(FilePathUtil.getResourceName(path), FilePathUtil.getResourcePath(path), response.size(), FilePathUtil.getType(path));
         } catch (ErrorResponseException e) {
             if (e.errorResponse().code().equals("NoSuchKey")) {
                 throw new ObjectNotFoundException("object not found");
@@ -105,7 +113,7 @@ public class FileService {
             if (e.errorResponse().code().equals("NoSuchKey")) {
                 throw new ObjectNotFoundException("object not found");
             }
-            throw new AppException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+            throw new RuntimeException();
         }
         catch (Exception e){
             throw new RuntimeException();
@@ -133,7 +141,7 @@ public class FileService {
                         if(FilePathUtil.getResourceName(item.objectName()).equals(query)){
                             infoDtos.add(new FileInfoDto(
                                     FilePathUtil.getResourceName(item.objectName()),
-                                    FilePathUtil.getPath(item.objectName()),
+                                    FilePathUtil.getResourcePath(item.objectName()),
                                     item.size(),
                                     FilePathUtil.getType(item.objectName())
                                     )
@@ -147,4 +155,85 @@ public class FileService {
 
         return infoDtos;
     }
+
+    public List<FileInfoDto> getFolderResources(String path){
+        if(!FilePathUtil.isValidPath(path) || FilePathUtil.getType(path).equals("FILE")){
+            throw new InvalidInputDataException("Invalid path");
+        }
+
+        List<FileInfoDto> infoDtos = new ArrayList<>();
+
+        Iterable<Result<Item>> results = minioClient.listObjects(
+                ListObjectsArgs.builder()
+                        .bucket(bucketName)
+                        .prefix(getUserPrefix() + path)
+                        .recursive(false)
+                        .build());
+
+        results.forEach(
+                itemResult -> {
+                    try {
+                        Item item = itemResult.get();
+                        infoDtos.add(new FileInfoDto(
+                                        FilePathUtil.getResourceName(item.objectName()),
+                                        FilePathUtil.getResourcePath(item.objectName()),
+                                        item.size(),
+                                        FilePathUtil.getType(item.objectName())
+                                )
+                        );
+                    } catch (Exception e) {
+                        throw new RuntimeException(e);
+                    }
+                }
+        );
+        return infoDtos;
+    }
+
+    public FileInfoDto createFolder(String path){
+        if(!FilePathUtil.isValidPath(path) || FilePathUtil.getResourceName(path).equals("FILE")){
+            throw new InvalidInputDataException("Invalid path");
+        }
+
+        if(isPathToResourceExists(path)){
+            throw new ObjectAlreadyExistsException("directory already exists");
+        }
+
+        if(!isPathToResourceExists(FilePathUtil.getResourcePath(path))){
+            throw new ObjectNotFoundException("Parent folder didn't exist");
+        }
+
+
+        try {
+            minioClient.putObject(
+                    PutObjectArgs.builder().bucket(bucketName).object(getUserPrefix() + path).stream(
+                                    new ByteArrayInputStream(new byte[] {}), 0, -1)
+                            .build());
+        }
+        catch (Exception e){
+            throw new RuntimeException();
+        }
+        return getInfo(path);
+    }
+
+
+    private Boolean isPathToResourceExists(String path){
+        try {
+            minioClient.getObject(
+                    GetObjectArgs.builder()
+                            .bucket(bucketName)
+                            .object(getUserPrefix() + path)
+                            .build());
+            return true;
+        } catch (ErrorResponseException e) {
+            if (e.errorResponse().code().equals("NoSuchKey")) {
+                return false;
+            }
+            throw new RuntimeException();
+        }
+        catch (Exception e){
+            throw new RuntimeException();
+        }
+    }
+
+
 }
